requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20211102.0=h6b3803e_1
appnope=0.1.4=pyhd8ed1ab_0
arrow-cpp=14.0.2=hc7aafb3_1
asttokens=2.4.1=pyhd8ed1ab_0
aws-c-auth=0.6.19=h80987f9_0
aws-c-cal=0.5.20=h80987f9_0
aws-c-common=0.8.5=h80987f9_0
aws-c-compression=0.2.16=h80987f9_0
aws-c-event-stream=0.2.15=h313beb8_0
aws-c-http=0.6.25=h80987f9_0
aws-c-io=0.13.10=h80987f9_0
aws-c-mqtt=0.7.13=h80987f9_0
aws-c-s3=0.1.51=h80987f9_0
aws-c-sdkutils=0.1.6=h80987f9_0
aws-checksums=0.1.13=h80987f9_0
aws-crt-cpp=0.18.16=h313beb8_0
aws-sdk-cpp=1.10.55=h313beb8_0
boost-cpp=1.84.0=hca5e981_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
bzip2=1.0.8=h93a5062_5
c-ares=1.26.0=h93a5062_0
ca-certificates=2024.2.2=hf0a4a13_0
certifi=2024.2.2=pyhd8ed1ab_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=py310hd137fd4_0
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.8.0=py310h1253130_1
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
fonttools=4.48.1=py310hd125d64_0
freetype=2.12.1=hadb7bae_2
gflags=2.2.2=hc88da5d_1004
glog=0.5.0=h5c6a83d_0
grpc-cpp=1.48.2=hc60591f_1
icu=73.2=hc8870d7_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
ipykernel=6.29.2=pyh3cd1d5f_0
ipython=8.21.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.1=py310hbe9552e_0
keithley2600=2.1.0=pypi_0
kiwisolver=1.4.5=py310h38f39d4_1
krb5=1.21.2=h92f50d5_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=21_osxarm64_openblas
libboost=1.84.0=h238a360_0
libboost-devel=1.84.0=hf450f58_0
libboost-headers=1.84.0=hce30654_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcblas=3.9.0=21_osxarm64_openblas
libcurl=8.5.0=h2d989ff_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.19=hb547adb_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libevent=2.1.12=h2757513_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=21_osxarm64_openblas
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.26=openmp_h6c19121_0
libpng=1.6.42=h091b4b1_0
libprotobuf=3.20.3=hb5ab8b9_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.44.2=h091b4b1_0
libssh2=1.11.0=h7a5bd25_0
libthrift=0.15.0=h73c2103_2
libtiff=4.6.0=ha8a6c65_2
libwebp-base=1.3.2=hb547adb_0
libxcb=1.15=hf346824_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=17.0.6=hcd81f8e_0
lz4-c=1.9.4=hb7217d7_0
matplotlib=3.8.2=py310hb6292c7_0
matplotlib-base=3.8.2=py310h9d2df84_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=h463b476_2
nest-asyncio=1.6.0=pyhd8ed1ab_0
numpy=1.26.4=py310hd45542a_0
openjpeg=2.5.0=h4c1507b_3
openssl=3.2.1=h0d3ecfb_0
orc=1.7.4=hdca1487_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.0=py310h401b61c_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=py310hfae7ebd_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
psutil=5.9.8=py310hd125d64_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=14.0.2=py310h68338f1_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyserial=3.5=pyhd8ed1ab_0
pysimplegui=4.60.5=pyhd8ed1ab_0
python=3.10.13=h2469fbe_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.4=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pytz=2024.1=pyhd8ed1ab_0
pyvisa=1.14.1=pypi_0
pyvisa-py=0.7.1=pypi_0
pyzmq=25.1.2=py310hbb13138_0
re2=2022.04.01=h6b3803e_0
readline=8.2=h92ec313_1
scipy=1.12.0=py310hf4b343e_2
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h17c5cce_0
stack_data=0.6.2=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tornado=6.3.3=py310h2aa6e3c_1
traitlets=5.14.1=pyhd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
unicodedata2=15.1.0=py310h2aa6e3c_0
utf8proc=2.6.1=h80987f9_1
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
zeromq=4.3.5=h965bd2d_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.5=h4f39d0f_0
